+++ src/main/java/io/github/sefiraat/networks/utils/datatypes/PersistentQuantumStorageType.java
@@ -47,7 +47,7 @@ public class PersistentQuantumStorageType implements PersistentDataType<Persiste
         final PersistentDataContainer container = context.newPersistentDataContainer();
 
         container.set(ITEM, DataType.ITEM_STACK, complex.getItemStack());
-        container.set(AMOUNT, DataType.INTEGER, complex.getAmount());
+        container.set(AMOUNT, DataType.LONG, complex.getAmount());
         container.set(MAX_AMOUNT, DataType.INTEGER, complex.getLimit());
         container.set(VOID, DataType.BOOLEAN, complex.isVoidExcess());
         return container;
